backend/requirements.txt
python-dotenv==1.0.0
uvicorn==0.23.1
fastapi==0.100.1
SQLAlchemy==1.4.49
databases==0.7.0
asyncpg==0.28.0
greenlet==2.0.2
python-multipart==0.0.6
bcrypt==4.0.1
passlib==1.7.4
python-jose==3.3.0

# Conciliación bancaria
openai==1.54.0
PyPDF2==3.0.1
pandas==2.3.3
openpyxl==3.1.2

# alembic==1.11.3
# annotated-types==0.5.0
# anyio==3.7.1
# asyncio==3.4.3
# asyncpg==0.28.0
# bcrypt==4.0.1
# click==8.1.6
# databases==0.7.0
# ecdsa==0.18.0
# exceptiongroup==1.1.2
# fastapi==0.100.1
# greenlet==2.0.2
# h11==0.14.0
# idna==3.4
# Mako==1.2.4
# MarkupSafe==2.1.3
# passlib==1.7.4
# psycopg2-binary==2.9.6
# pyasn1==0.5.0
# pydantic==2.1.1
# pydantic_core==2.4.0
# python-dotenv==1.0.0
# python-jose==3.3.0
# python-multipart==0.0.6
# rsa==4.9
# six==1.16.0
# sniffio==1.3.0
# SQLAlchemy==1.4.49
# starlette==0.27.0
# typing_extensions==4.7.1
# uvicorn==0.23.1
